Begin renaming process for Recipe(s) and Target
