+++ lib/revolution/recipes.rb
